rename test_single to test and add support for multicams combine OffsetSlider to utils/offsets.py
